Advertisements

+++ InvestigationProject/app/Controller/PublicationController.php
@@ -5,7 +5,8 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-App::uses('MembersSection', 'Model');
+App::uses('Section', 'Model');
+App::uses('PublicationsSectionField', 'Model');
 
 /**
  * Description of PublicationController
@@ -42,16 +43,15 @@ class PublicationController extends AppController {
 
     public function register() {
         $this->set('page_name', 'Registrar publicación');
-        $member_section_db = new MembersSection();
-        $section_options = $member_section_db->find('list', array(
-            'fields' => array('Section.id', 'Section.name'),
-            'conditions' => array('MembersSection.member_id' => $this->Session->read('User.member_id')),
+        $section_db = new Section();
+        $section_options = $section_db->find('list', array(
+            'fields' => array('Section.id', 'Section.name'),            
             'recursive' => 1
         ));
         $this->set('section_options', $section_options);
         if ($this->request->is('post')) {
             if (!empty($this->data)) {
-                $this->request->data['Publication']['member_id'] = $this->Session->read('User.member_id');
+                $this->request->data['Publication']['member_id'] = $this->Session->read('User.member_id');                                
                 if (is_uploaded_file($this->request->data['Publication']['file_path']['tmp_name'])) {
                     $filename = basename($this->request->data['Publication']['file_path']['name']);
                     $path = WWW_ROOT . DS . 'files' . DS . 'publications' . DS . 'member_' . $this->Session->read('User.member_id') . DS;
@@ -60,17 +60,21 @@ class PublicationController extends AppController {
                     }
                     move_uploaded_file(
                             $this->data['Publication']['file_path']['tmp_name'], $path . $filename
-                    );
+                    );                    
                     $this->request->data['Publication']['file_path'] = $path . $filename;
-                    if ($this->Publication->saveAll($this->request->data)) {
-                        $this->Session->setFlash('Se ha creado la publicación ' . $this->data['Publication']['title'], 'success-message');
-                        return $this->redirect(array(
-                                    'controller' => 'publication',
-                                    'action' => 'index',
-                                    'mine'));
-                    } else {
+                } else {
+                    $this->request->data['Publication']['file_path'] = '';
+                }                
+                if ($this->Publication->saveAll($this->request->data)) {
+                    $this->Session->setFlash('Se ha creado la publicación ' . $this->data['Publication']['title'], 'success-message');
+                    return $this->redirect(array(
+                                'controller' => 'publication',
+                                'action' => 'index',
+                                'mine'));
+                } else {
+                    $this->Session->setFlash('Ocurrió un error al guardar la publicación ' . $this->data['Publication']['title'], 'alert-message');
+                    if(isset($path)) {
                         unlink($path . $filename);
-                        $this->Session->setFlash('Ocurrió un error al guardar la publicación ' . $this->data['Publication']['title'], 'error-message');
                     }
                 }
             }
